refactor operations
