+++ base/src/main/java/com/smartdevicelink/managers/screen/choiceset/BaseChoiceSetManager.java
@@ -39,12 +39,6 @@ import android.support.annotation.NonNull;
 import com.smartdevicelink.managers.BaseSubManager;
 import com.smartdevicelink.managers.CompletionListener;
 import com.smartdevicelink.managers.file.FileManager;
-import com.smartdevicelink.managers.screen.choiceset.operations.CheckChoiceVROptionalInterface;
-import com.smartdevicelink.managers.screen.choiceset.operations.CheckChoiceVROptionalOperation;
-import com.smartdevicelink.managers.screen.choiceset.operations.DeleteChoicesOperation;
-import com.smartdevicelink.managers.screen.choiceset.operations.PreloadChoicesOperation;
-import com.smartdevicelink.managers.screen.choiceset.operations.PresentChoiceSetOperation;
-import com.smartdevicelink.managers.screen.choiceset.operations.PresentKeyboardOperation;
 import com.smartdevicelink.protocol.enums.FunctionID;
 import com.smartdevicelink.proxy.RPCNotification;
 import com.smartdevicelink.proxy.interfaces.ISdl;
