+++ base/src/main/java/com/smartdevicelink/managers/screen/choiceset/CheckChoiceVROptionalInterface.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019 Livio, Inc.
+ * Copyright (c)  2019 Livio, Inc.
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -29,23 +29,13 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  *
- * Created by brettywhite on 2019-06-11.
+ * Created by brettywhite on 6/12/19 1:52 PM
+ *
  */
 
-package com.smartdevicelink.managers.screen.choiceset.operations;
+package com.smartdevicelink.managers.screen.choiceset;
 
-import com.smartdevicelink.AndroidTestCase2;
-
-public class PresentKeyboardOperationTests extends AndroidTestCase2 {
-
-	@Override
-	public void setUp() throws Exception{
-		super.setUp();
-	}
-
-	@Override
-	public void tearDown() throws Exception {
-		super.tearDown();
-	}
-
+public interface CheckChoiceVROptionalInterface {
+	void onCheckChoiceVROperationComplete(boolean vrOptional);
+	void onError(String error);
 }
