+++ base/src/main/java/com/smartdevicelink/managers/screen/choiceset/PreloadChoicesOperation.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019 Livio, Inc.
+ * Copyright (c)  2019 Livio, Inc.
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -29,10 +29,11 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  *
- * Created by brettywhite on 2019-06-11.
+ * Created by brettywhite on 6/12/19 1:52 PM
+ *
  */
 
-package com.smartdevicelink.managers.screen.choiceset.operations;
+package com.smartdevicelink.managers.screen.choiceset;
 
 import android.support.annotation.NonNull;
